// src/xpanse-api/generated/models/RegisterServiceEntity.ts
/*
 * SPDX-License-Identifier: Apache-2.0
 * SPDX-FileCopyrightText: Huawei Inc.
 */

/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Ocl } from './Ocl';

export class RegisterServiceEntity {
    'createTime'?: Date;
    'lastModifiedTime'?: Date;
    'id'?: string;
    'name'?: string;
    'version'?: string;
    'csp'?: RegisterServiceEntityCspEnum;
    'category'?: RegisterServiceEntityCategoryEnum;
    'ocl'?: Ocl;
    'serviceState'?: RegisterServiceEntityServiceStateEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "createTime",
            "baseName": "createTime",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "lastModifiedTime",
            "baseName": "lastModifiedTime",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "version",
            "baseName": "version",
            "type": "string",
            "format": ""
        },
        {
            "name": "csp",
            "baseName": "csp",
            "type": "RegisterServiceEntityCspEnum",
            "format": ""
        },
        {
            "name": "category",
            "baseName": "category",
            "type": "RegisterServiceEntityCategoryEnum",
            "format": ""
        },
        {
            "name": "ocl",
            "baseName": "ocl",
            "type": "Ocl",
            "format": ""
        },
        {
            "name": "serviceState",
            "baseName": "serviceState",
            "type": "RegisterServiceEntityServiceStateEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return RegisterServiceEntity.attributeTypeMap;
    }

    public constructor() {
    }
}


export type RegisterServiceEntityCspEnum = "aws" | "azure" | "alibaba" | "huawei" | "openstack" ;
export type RegisterServiceEntityCategoryEnum = "ai" | "compute" | "container" | "storage" | "network" | "database" | "media_service" | "security" | "middleware" | "others" ;
export type RegisterServiceEntityServiceStateEnum = "REGISTERED" | "UPDATED" | "DEPLOYING" | "DEPLOY_SUCCESS" | "DEPLOY_FAILED" | "DESTROYING" | "DESTROY_SUCCESS" | "DESTROY_FAILED" ;